+++ CHANGELOG.md
@@ -12,6 +12,11 @@ changes that do not affect the user.
 
 - Added Python 3.13 classifier in pyproject.toml (we now also run tests on Python 3.13 in the CI).
 
+### Added
+
+- New aggregator `ConFIG` from [ConFIG: Towards Conflict-free Training of Physics
+Informed Neural Networks](https://arxiv.org/pdf/2408.11104).
+
 ## [0.4.1] - 2025-01-02
 
 ### Fixed